Drop ssh module
Preparation for switching to the `ssh-key` crate from `sshkeys`,
which will allow SSH key generation.
