+++ src/main.rs
@@ -72,7 +72,6 @@ mod docs_man;
 mod docs_markdown;
 mod iostreams;
 mod prompt_ext;
-mod ssh;
 mod types;
 
 #[cfg(test)]
